+++ prov/util/src/uffd_mem_monitor.c
@@ -181,12 +181,12 @@ static void ofi_uffd_pagefault_handler(struct uffd_msg *msg)
 	if (flags != UFFD_PAGEFAULT_FLAG_WRITE) {
 #if HAVE_UFFD_THREAD_ID
 		FI_WARN(&core_prov, FI_LOG_MR,
-			"UFFD pagefault with unrecognized flags: %lu, address %p, thread %u\n",
-			flags, address, ptid);
+			"UFFD pagefault with unrecognized flags: %" PRIu64 ", address %p, thread %u\n",
+			(uint64_t)flags, address, ptid);
 #else
 		FI_WARN(&core_prov, FI_LOG_MR,
-			"UFFD pagefault with unrecognized flags: %lu, address %p\n",
-			flags, address);
+			"UFFD pagefault with unrecognized flags: %" PRIu64 ", address %p\n",
+			(uint64_t)flags, address);
 #endif
 		/* The faulting thread is halted at this point. In
 		 * theory we could wake it up with UFFDIO_WAKE. In
